+++ lib/src/db/lib_prod_dbs.dart
@@ -18,6 +18,7 @@ class DbProd extends DbEnv {
   static final mis = Db('mongodb://${DbProd._connection}/mis');
   static final nyiso = Db('mongodb://${DbProd._connection}/nyiso');
   static final pjm = Db('mongodb://${DbProd._connection}/pjm');
+  static final retailSuppliers = Db('mongodb://${DbProd._connection}/retail_suppliers');
   static final riskSystem = Db('mongodb://${DbProd._connection}/risk_system');
   static final weather = Db('mongodb://${DbProd._connection}/weather');
 }